requirements.txt
requests
beautifulsoup4
trafilatura
curl_cffi
flask
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
